requirements.txt
uvicorn
fastapi

PyJWT
python-decouple
pydantic[email]
